Restructuring for searching
There's a number of changes, but they are mostly for the purpose of having a "normalized" field representing the text in a easily-searchable way. All it is is a search field with the string lowercased with spaces removed.

I'm not sure if removing the spaces is actually beneficial, and I might revert it depending on how it works in practice. I don't currently have the mental foresight to determine how it will turn out.